Move request_with_pycurl to utils
